+++ frontend/src/lib/components/notification_view/NotificationToolbar.svelte
@@ -32,7 +32,6 @@
 	export let onClear: () => void = () => {};
 	export let onUpdateView: () => void = () => {};
 	export let onSaveAsNewView: () => void = () => {};
-	export let saveNote: string | null = null;
 
 	// Pagination props
 	export let page: number;
@@ -138,7 +137,6 @@
 				{canUpdateView}
 				onSaveAsNew={onSaveAsNewView}
 				{onUpdateView}
-				{saveNote}
 			/>
 		</div>
 	{/if}