+++ frontend/src/lib/components/notification_view/NotificationView.svelte
@@ -46,7 +46,6 @@
 	export let onClear: () => void;
 	export let onUpdateView: () => void;
 	export let onSaveAsNewView: () => void;
-	export let saveNote: string | null = null;
 	export let page: number;
 	export let totalPages: number;
 	export let onPrevious: () => void;
@@ -222,7 +221,6 @@
 				{onClear}
 				{onUpdateView}
 				{onSaveAsNewView}
-				{saveNote}
 				{page}
 				{totalPages}
 				{onPrevious}